Added bot files
These are the 2 different bots that I have currently. 

I just coded them in Python with the help of ChatGPT4.

I'm sure there are a million ways these are broken and inefficient, but they mostly work fine and are seemingly pretty efficient on token usage considering how much context they can retain and react to.